Moved the editable flag from TDCContent to UserContent
