+++ src/main/java/org/tndata/android/compass/model/Action.java
@@ -84,5 +84,4 @@ public abstract class Action extends UserContent implements Serializable{
     public abstract String getTitle();
     public abstract Goal getGoal();
     public abstract String getGoalTitle();
-    public abstract boolean isEditable();
 }
