+++ src/main/java/org/tndata/android/compass/model/UserAction.java
@@ -151,10 +151,6 @@ public class UserAction
         return TYPE;
     }
 
-    @Override
-    public boolean isEditable(){
-        return mAction.isEditable();
-    }
 
     /*---------*
      * UTILITY *
