+++ src/main/java/org/tndata/android/compass/model/UserBehavior.java
@@ -119,11 +119,6 @@ public class UserBehavior
     }
 
     @Override
-    public boolean isEditable(){
-        return mBehavior.isEditable();
-    }
-
-    @Override
     protected String getType(){
         return TYPE;
     }
