+++ src/main/java/org/tndata/android/compass/model/UserCategory.java
@@ -86,11 +86,6 @@ public class UserCategory extends UserContent implements Serializable, UserSelec
     }
 
     @Override
-    public boolean isEditable(){
-        return mCategory.isEditable();
-    }
-
-    @Override
     protected String getType(){
         return TYPE;
     }
