+++ src/main/java/org/tndata/android/compass/model/UserGoal.java
@@ -127,11 +127,6 @@ public class UserGoal extends Goal implements Serializable, UserSelectedContent,
     }
 
     @Override
-    public boolean isEditable(){
-        return mGoal.isEditable();
-    }
-
-    @Override
     protected String getType(){
         return TYPE;
     }
