+++ src/main/java/org/tndata/android/compass/model/UserSelectedContent.java
@@ -10,7 +10,6 @@ public interface UserSelectedContent{
     String getDescription();
     String getHTMLDescription();
     String getIconUrl();
-    boolean isEditable();
 
     void init();
 }
